Add Open Graph meta tags for shared project links (#234)
Enable rich link previews on WhatsApp, Facebook, Twitter, and Google
when sharing project URLs. Splits the shared page into a server
component (for generateMetadata) and a client component (for the
interactive 3D viewer). Includes a default OG image fallback.
